add more configuration to go_proto_compile's protoc call
This adds the ability to configure how go_proto_compile wil call its configured protoc command.

The _protoc option is changed to protoc, to allow setting it externally.
The prefix_args_option option is added, to configure additional flags/args to pass to the protoc command before the derived flags and args.
The transitive_sources_option is added to pass the protoc command transitive source paths instead of binary filedescriptorsets.
The input_source_paths_option is added to pass the protoc command its inputs as their actual file paths, instead of as the import paths of those files.