proto: add extra 'go_srcs' option
This adds a 'go_srcs' attribute to the go_proto_library rule which
can be used to specify additional go source files that should be included
in the sources of the derived go_library. These can be required when using
non-proto-defined types in the generated proto, e.g. via 'casttype' which
are defined in files that also then reference proto-defined types,
creating a circular dependency between those files, which are intended
to be built as a single package. Thus those extra sources need to be
included in the go_proto_library's generated go_library.