sdk: symlink everything under lib/ when using local SDK
We rely on the packaged zoneinfo zip in CRDB, found under
lib/time/zoneinfo.zip. When using the remote SDK, we untar the entire
thing into our sandbox. When pointing to a local SDK however, we ignored
everything under lib, so failed to build.

We'll need this diff in order to build CRDB with modified go
installations, or at least iterating on it locally.

+cc https://github.com/cockroachdb/cockroach/issues/56178